requirements.txt
alembic
amqp
appdirs
asgiref
attrs
autopep8
beautifulsoup4
billiard
black
cached-property
cachetools
celery
certifi
cffi
chardet
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
defusedxml
distlib
Django
django-allauth
django-crispy-forms
et-xmlfile
filelock
Flask
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
forex-python
future
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleads
googleapis-common-protos
googlemaps
h2o
httplib2
idna
isodate
itsdangerous
Jinja2
joblib
jsonpickle
kombu
lxml
Mako
MarkupSafe
mypy-extensions
nltk
numerize
numpy
oauth2client
oauthlib
openpyxl
packaging
pandas
pathspec
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pycodestyle
pycparser
PyJWT
pyparsing
pyseoanalyzer
python-dateutil
python-dotenv
python-editor
python-social-auth
python3-openid
pytz
PyYAML
redis
regex
requests
requests-file
requests-oauthlib
requests-toolbelt
rq
rsa
simplejson
six
social-auth-app-django
social-auth-core
soupsieve
SQLAlchemy
sqlparse
stop-words
tabulate
toml
tqdm
typed-ast
typing-extensions
uritemplate
urllib3
vine
virtualenv
wcwidth
Werkzeug
xlrd
xmltodict
zeep
google-auth-oauthlib
psycopg2-binary
platformdirs
selenium
tomli
appnope
async-generator
Babel
backcall
bcrypt
blinker
cycler
debugpy
decorator
django-session-timeout
entrypoints
extras
fixtures
Flask-BabelEx
Flask-Gravatar
Flask-HTMLmin
Flask-Login
Flask-Mail
Flask-Paranoid
Flask-Principal
Flask-Security
Flask-WTF
gunicorn
h11
html5lib
htmlmin
install
ipykernel
ipython
ipython-genutils
jedi
jupyter-client
jupyter-core
kiwisolver
linecache2
matplotlib
matplotlib-inline
nest-asyncio
outcome
paramiko
parso
passlib
pbr
pexpect
pickleshare
Pillow
ptyprocess
py4j
Pygments
PyNaCl
pyOpenSSL
pyrsistent
pyspark
python-mimeparse
pyzmq
scipy
seaborn
sniffio
sortedcontainers
speaklater
sshtunnel
testtools
tornado
traceback2
traitlets
trio
trio-websocket
unittest2
webencodings
wsproto
WTForms
